backend/requirements.txt
django==4.2.16
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.4.0
setuptools>=70.0.0